Addition to previous commit and bugfix.
Previously, if a file extension was set to be opened externally, and
the external program was Atom, an endless loop would occur. A hacky fix for
this has been added by the way of a hashtable to only allow a certain
amount of tries for a given path.

Added new setting to allow user to specify that protocol part of URI
should be disregarded. See issue #5 for more information.